+++ src/Symfony/Component/DependencyInjection/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 7.1
 ---
 
+ * Add method `date` prefix to `EnvVarProcessor` that allows get a DateTimeImmutable from an env var
  * Add argument `$prepend` to `ContainerConfigurator::extension()` to prepend the configuration instead of appending it
 
 7.0